Core Sync: 3179: Allow null enchant routes. 3181: Unhardcoding message ids in DlgAnswer packet. 3182: Sync resurrect skill with dp(handlers) 3184: Fix for arrow glitch 3185: Support for new arrows/bolts 3186: Fix system message when applying dye 3187: Fix some system message and attribute on SLOT_BACK (thx rocknow) 3188: Two little performance tuning (List<Integer> -> int[] with binarySearch) 3189: If registered players die in Castle siege zone then they get Battlefield Death Syndrome. 3190: Elemental system formula update. Also using byte data type for elements.
git-svn-id: http://svn.l2jfree.com/l2jfree@6408 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/ItemInfo.java
@@ -57,7 +57,7 @@ public class ItemInfo implements ElementalOwner
 
 	private int _location;
 
-	private int _elemAtkType = -2;
+	private byte _elemAtkType = -2;
 	private int _elemAtkPower = 0;
 	private int[] _elemDefAttr = {0, 0, 0, 0, 0, 0};
 
@@ -233,7 +233,7 @@ public class ItemInfo implements ElementalOwner
 		return _location;
 	}
 
-	public int getAttackElementType()
+	public byte getAttackElementType()
 	{
 		return _elemAtkType;
 	}
